feat: Support bundled JS applications
Instead of parsing and statically looking for module exports in the
compile step, this evaluates the code to find the exports. Thus we can
support code bundled by esbuild or webpack. Granted, right now we can
only support CJS as a target because quickjs doesn't seem to like the
export keyword.