fix: uncomment python
I suspect the `extism_sys` errors might have been a red herring.
It looks like CI was failing trying to find the engine. This commit
should give us more information about the CI job. If you see this
commit on main: that's confirmation that the `extism_sys` bug was
a local issue.